+++ test/command_tester.sh
@@ -2,10 +2,6 @@
 
 echo "Script to test the luajit Terminal Functions"
 
-# Build First
-echo "Build the executable First"
-cargo build
-
 echo "Checking Version:"
 ./target/debug/luajit --version