+++ CHANGELOG.md
@@ -15,6 +15,9 @@
 - added Script to get all versions
 - reformatted some folders
 - reformatted some scripts
+- added option to give arguments to luajit
+- added Client to the Download File function
+- added Colors to the help Info
 
 ## 0.1.12 - 08.10.2025
 - added yaml parser